GH-12103 refactor product config e2e helpers and tests
