+++ projects/storefrontapp-e2e-cypress/cypress/integration/mobile/product-configuration-mobile.flaky-e2e-spec.ts
@@ -1,5 +1,5 @@
-import * as configurationCommon from '../../helpers/product-configuration';
-import * as configuration from '../../helpers/product-configurator-vc';
+import * as configuration from '../../helpers/product-configuration';
+import * as configurationVc from '../../helpers/product-configurator-vc';
 import { formats } from '../../sample-data/viewports';
 
 /**
@@ -26,14 +26,14 @@ context('Product Configuration', () => {
     it('should navigate using the group menu in mobile resolution', () => {
       cy.window().then((win) => win.sessionStorage.clear());
       cy.viewport(formats.mobile.width, formats.mobile.height);
-      configuration.goToConfigurationPage(electronicsShop, testProduct);
-      configurationCommon.checkHamburgerDisplayed();
+      configurationVc.goToConfigurationPage(electronicsShop, testProduct);
+      configuration.checkHamburgerDisplayed();
       configuration.checkAttributeDisplayed(CAMERA_MODE, radioGroup);
 
       configuration.clickHamburger();
       configuration.checkGroupMenuDisplayed();
 
-      configuration.clickOnGroup(2);
+      configurationVc.clickOnGroup(2);
       configuration.checkAttributeDisplayed(CAMERA_DISPLAY, radioGroup);
     });
   });
