+++ projects/storefrontapp-e2e-cypress/cypress/integration/regression/product-configuration-cart.flaky-e2e-spec.ts
@@ -1,9 +1,11 @@
 import * as cart from '../../helpers/cart';
 import * as login from '../../helpers/login';
-import * as configurationCommon from '../../helpers/product-configuration';
-import * as configurationOverviewCommon from '../../helpers/product-configuration-overview';
-import * as configuration from '../../helpers/product-configurator-vc';
-import * as configurationOverview from '../../helpers/product-configurator-vc-overview';
+import * as configuration from '../../helpers/product-configuration';
+import * as configurationOverview from '../../helpers/product-configuration-overview';
+import * as configurationVc from '../../helpers/product-configurator-vc';
+import * as configurationOverviewVc from '../../helpers/product-configurator-vc-overview';
+import * as configurationCart from '../../helpers/product-configurator-cart';
+import * as configurationCartVc from '../../helpers/product-configurator-cart-vc';
 import * as productSearch from '../../helpers/product-search';
 
 /**
@@ -55,14 +57,14 @@ context('Product Configuration', () => {
 
   describe('Navigate to Product Configuration Page', () => {
     it('should be able to navigate from the cart', () => {
-      configuration.goToConfigurationPage(
+      configurationVc.goToConfigurationPage(
         electronicsShop,
         testProductMultiLevel
       );
-      configuration.clickAddToCartBtn();
-      configuration.goToCart(electronicsShop);
+      configurationVc.clickAddToCartBtn();
+      configurationVc.goToCart(electronicsShop);
       //We assume only one product is in the cart
-      configuration.clickOnEditConfigurationLink(0);
+      configurationCart.clickOnEditConfigurationLink(0);
     });
 
     it('should be able to navigate from the cart after adding product directly to the cart', () => {
@@ -75,10 +77,10 @@ context('Product Configuration', () => {
       ).as('productSearch');
       productSearch.searchForProduct(testProductMultiLevel);
       cy.wait('@productSearch');
-      configurationCommon.clickOnAddToCartBtnOnPD();
-      configurationCommon.clickOnViewCartBtnOnPD();
+      configuration.clickOnAddToCartBtnOnPD();
+      configuration.clickOnViewCartBtnOnPD();
       cart.verifyCartNotEmpty();
-      configuration.clickOnEditConfigurationLink(0);
+      configurationCart.clickOnEditConfigurationLink(0);
     });
   });
 
@@ -91,7 +93,7 @@ context('Product Configuration', () => {
           'BASE_SITE'
         )}/ccpconfigurator/*`
       ).as('updateConfig');
-      configuration.goToConfigurationPage(
+      configurationVc.goToConfigurationPage(
         electronicsShop,
         testProductMultiLevel
       );
@@ -99,26 +101,26 @@ context('Product Configuration', () => {
       configuration.selectAttribute(PROJECTOR_TYPE, radioGroup, PROJECTOR_LCD);
       cy.wait('@updateConfig');
       configuration.clickOnPreviousBtn(GENERAL);
-      configuration.clickOnGroup(3);
+      configurationVc.clickOnGroup(3);
 
-      configuration.selectConflictingValue(
+      configurationVc.selectConflictingValue(
         GAMING_CONSOLE,
         radioGroup,
         GAMING_CONSOLE_YES,
         1
       );
       cy.wait('@updateConfig');
-      configuration.checkStatusIconDisplayed(SOURCE_COMPONENTS, WARNING);
-      configuration.checkStatusIconDisplayed(VIDEO_SYSTEM, WARNING);
-      configuration.deselectConflictingValue(
+      configurationVc.checkStatusIconDisplayed(SOURCE_COMPONENTS, WARNING);
+      configurationVc.checkStatusIconDisplayed(VIDEO_SYSTEM, WARNING);
+      configurationVc.deselectConflictingValue(
         GAMING_CONSOLE,
         radioGroup,
         GAMING_CONSOLE_NO
       );
       cy.wait('@updateConfig');
-      configuration.checkStatusIconNotDisplayed(SOURCE_COMPONENTS);
-      configuration.checkStatusIconNotDisplayed(VIDEO_SYSTEM);
-      configuration.selectConflictingValue(
+      configurationVc.checkStatusIconNotDisplayed(SOURCE_COMPONENTS);
+      configurationVc.checkStatusIconNotDisplayed(VIDEO_SYSTEM);
+      configurationVc.selectConflictingValue(
         GAMING_CONSOLE,
         radioGroup,
         GAMING_CONSOLE_YES,
@@ -131,46 +133,46 @@ context('Product Configuration', () => {
       configuration.clickOnPreviousBtn(FRONT_SPEAKERS);
       configuration.clickOnPreviousBtn(PROJECTOR_SCREEN);
       configuration.clickOnPreviousBtn(PROJECTOR);
-      configuration.checkConflictDetectedMsgDisplayed(PROJECTOR_TYPE);
+      configurationVc.checkConflictDetectedMsgDisplayed(PROJECTOR_TYPE);
       configuration.clickOnPreviousBtn(GENERAL);
       configuration.clickOnPreviousBtn(CONFLICT_FOR_GAMING_CONSOLE);
-      configuration.checkConflictDescriptionDisplayed(
+      configurationVc.checkConflictDescriptionDisplayed(
         Conflict_msg_gaming_console
       );
       configuration.clickOnNextBtn(GENERAL);
-      configuration.checkStatusIconDisplayed(SOURCE_COMPONENTS, WARNING);
-      configuration.checkStatusIconDisplayed(VIDEO_SYSTEM, WARNING);
-      configurationOverviewCommon.registerConfigurationOvOCC();
-      configuration.clickAddToCartBtn();
+      configurationVc.checkStatusIconDisplayed(SOURCE_COMPONENTS, WARNING);
+      configurationVc.checkStatusIconDisplayed(VIDEO_SYSTEM, WARNING);
+      configurationOverviewVc.registerConfigurationOvOCC();
+      configurationVc.clickAddToCartBtn();
       // Navigate to Overview page and verify whether the resolve issues banner is displayed and how many issues are there
-      configurationOverview.verifyNotificationBannerOnOP(1);
+      configurationOverviewVc.verifyNotificationBannerOnOP(1);
       // Navigate to cart and verify whether the  the resolve issues banner is displayed and how many issues are there
-      configurationOverviewCommon.clickContinueToCartBtnOnOP();
-      configuration.verifyNotificationBannerInCart(0, 1);
+      configurationOverview.clickContinueToCartBtnOnOP();
+      configurationCartVc.verifyNotificationBannerInCart(0, 1);
       // Navigate back to the configuration page
-      configuration.clickOnEditConfigurationLink(0);
+      configurationCart.clickOnEditConfigurationLink(0);
       // Navigate to Overview page and back to configuration via 'Resolve issues' link
-      configuration.clickAddToCartBtn();
+      configurationVc.clickAddToCartBtn();
       // Click 'Resolve issues' link in the banner and navigate back to the configuration
-      configurationOverview.clickOnResolveIssuesLinkOnOP();
-      configuration.checkConflictDescriptionDisplayed(
+      configurationOverviewVc.clickOnResolveIssuesLinkOnOP();
+      configurationVc.checkConflictDescriptionDisplayed(
         Conflict_msg_gaming_console
       );
       configuration.clickOnNextBtn(GENERAL);
       // Navigate back to the configuration page and deselect conflicting value
-      configuration.clickOnGroup(3);
-      configuration.deselectConflictingValue(
+      configurationVc.clickOnGroup(3);
+      configurationVc.deselectConflictingValue(
         GAMING_CONSOLE,
         radioGroup,
         GAMING_CONSOLE_NO
       );
       //Click 'Add to cart' and verify whether the resolve issues banner is not displayed anymore
-      configurationOverviewCommon.registerConfigurationOvOCC();
-      configuration.clickAddToCartBtn();
-      configurationOverview.verifyNotificationBannerOnOP();
+      configurationOverviewVc.registerConfigurationOvOCC();
+      configurationVc.clickAddToCartBtn();
+      configurationOverviewVc.verifyNotificationBannerOnOP();
       // Click 'Continue to cart' and verify whether there is a resolve issues banner in the cart entry list
-      configurationOverviewCommon.clickContinueToCartBtnOnOP();
-      configuration.verifyNotificationBannerInCart(0);
+      configurationOverview.clickContinueToCartBtnOnOP();
+      configurationCartVc.verifyNotificationBannerInCart(0);
     });
   });
 
@@ -179,10 +181,10 @@ context('Product Configuration', () => {
       login.registerUser();
       login.loginUser();
       productSearch.searchForProduct(testProductMultiLevel);
-      configurationCommon.clickOnAddToCartBtnOnPD();
-      configurationCommon.clickOnProceedToCheckoutBtnOnPD();
-      configurationCommon.checkout();
-      configurationCommon.navigateToOrderDetails();
+      configuration.clickOnAddToCartBtnOnPD();
+      configuration.clickOnProceedToCheckoutBtnOnPD();
+      configurationCartVc.checkout();
+      configurationCart.navigateToOrderDetails();
       //don't check the order history aspect because this part is flaky
       //configuration.selectOrderByOrderNumberAlias();
       login.signOutUser();
